+++ src/transpilation/index.ts
@@ -1,4 +1,6 @@
-export * from "./managed";
+export { Chunk } from "./chunk";
+export * from "./managed-api";
+export { Module } from "./module";
 export { Plugin } from "./plugins";
 export * from "./transpile";
 export * from "./transpiler";
